Explicitly cast to silence a compiler warning.
As of the time of this writing max_fat is generally around 25; no way it’ll overflow.

+++ main.m
@@ -1091,7 +1091,7 @@ void printExecutableArchitectures(CFURLRef url, bool printOnFailure) {
 
     // Look for any of the six magic numbers relevant to Mach-O executables, and swap the header if necessary.
     uint32_t num_fat = 0, magic = *((uint32_t *)bytes);
-    uint32_t max_fat = (length - sizeof(struct fat_header)) / sizeof(struct fat_arch);
+    uint32_t max_fat = (uint32_t)(length - sizeof(struct fat_header)) / sizeof(struct fat_arch);
     struct fat_arch one_fat = {0}, *fat;
     if (MH_MAGIC == magic || MH_CIGAM == magic) {
         struct mach_header *mh = (struct mach_header *)bytes;
